requirements.txt
mysql-connector-python==8.0.29
numpy==1.22.4
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
pandas==1.4.2
Pillow==9.1.1
tk==
tkinter